requirements.txt
Django==1.10.3
django-model-utils==2.6
